Fix compiled functions when # of upvalues >= 128

+++ luaj-vm/src/core/org/luaj/compiler/FuncState.java
@@ -131,8 +131,8 @@ public class FuncState extends LuaC {
 		f.upvalues[f.nups] = name;
 		_assert (v.k == LexState.VLOCAL || v.k == LexState.VUPVAL);
 		upvalues[f.nups] = new upvaldesc();
-		upvalues[f.nups].k = (byte) (v.k);
-		upvalues[f.nups].info = (byte) (v.u.s.info);
+		upvalues[f.nups].k = (short) (v.k);
+		upvalues[f.nups].info = (short) (v.u.s.info);
 		return f.nups++;
 	}
